Move take, factorize
